+++ src/cpp_bindings/dense.cpp
@@ -6,6 +6,7 @@
 
 void init_dense(py::module_ &module_matrix)
 {
+    using dim_type = gko::dim<2>::dimension_type;
     /* function to create a dense matrix from py::buffer object
      *
      */
@@ -27,7 +28,10 @@ void init_dense(py::module_ &module_matrix)
         auto cols = (info.ndim == 1) ? 1 : info.shape[1];
 
         return gko::matrix::Dense<ValueType>::create(
-            exec, gko::dim<2>{rows, cols}, view, cols);
+            exec,
+            gko::dim<2>{static_cast<dim_type>(rows),
+                        static_cast<dim_type>(cols)},
+            view, cols);
     };
 
     py::class_<gko::matrix::Dense<ValueType>,
@@ -47,12 +51,13 @@ void init_dense(py::module_ &module_matrix)
                 exec,
                 gko::dim<2>{dim[0].cast<size_t>(), dim[1].cast<size_t>()}));
         }))
-        .def(py::init(
-            [](std::shared_ptr<gko::Executor> exec, py::tuple dim, int stride) {
-                return gko::share(gko::matrix::Dense<ValueType>::create(
-                    exec, gko::dim<2>{dim[0].cast<int>(), dim[1].cast<int>()},
-                    stride));
-            }))
+        .def(py::init([](std::shared_ptr<gko::Executor> exec, py::tuple dim,
+                         int stride) {
+            return gko::share(gko::matrix::Dense<ValueType>::create(
+                exec,
+                gko::dim<2>{dim[0].cast<dim_type>(), dim[1].cast<dim_type>()},
+                stride));
+        }))
         .def(py::init([](std::shared_ptr<gko::Executor> exec, py::tuple dim,
                          py::buffer b, size_t stride) {
             py::buffer_info info = b.request();
@@ -72,13 +77,15 @@ void init_dense(py::module_ &module_matrix)
             auto cols = (info.ndim == 1) ? 1 : info.shape[1];
 
             return gko::share(gko::matrix::Dense<ValueType>::create(
-                exec, gko::dim<2>{dim[0].cast<size_t>(), dim[1].cast<size_t>()},
+                exec,
+                gko::dim<2>{dim[0].cast<dim_type>(), dim[1].cast<dim_type>()},
                 view, stride));
         }))
         .def(py::init([](std::shared_ptr<gko::Executor> exec, py::tuple dim,
                          gko::array<ValueType> view, size_t stride) {
             return gko::share(gko::matrix::Dense<ValueType>::create(
-                exec, gko::dim<2>{dim[0].cast<size_t>(), dim[1].cast<size_t>()},
+                exec,
+                gko::dim<2>{dim[0].cast<dim_type>(), dim[1].cast<dim_type>()},
                 view, stride));
         }))
         .def("__repr__",