+++ src/pyGinkgo/core.py
@@ -4,6 +4,7 @@
 
 from pyGinkgo import pyGinkgoBindings as pGB
 import numpy as np
+import json
 
 
 def asarray(obj, executor: str = "Reference", dtype="float"):
@@ -27,28 +28,39 @@ def asarray(obj, executor: str = "Reference", dtype="float"):
     return ctr(executor, obj)
 
 
-def solve(A, b, initial_guess=None, solver="GMRES", solver_args=dict(): dict):
+def solve(A, b, initial_guess=None, solver_args: dict = dict()):
     """Solve a given linear system, where A is the system matrix and b the RHS
 
     Parameters: A - The system matrix
                 b - The right hand side vector
-                initial_guess - The initial guess 
+                initial_guess - The initial guess
                 solver - The solver
                 solver_args - A dictionary that is forwarded to the solver containing
-                    arguments, eg {'max_iters': 100, 'tolerance': 1e-6}  
+                    arguments, eg {'max_iters': 100, 'tolerance': 1e-6}
+    Returns: tuple of a logger object and solution vector
     """
 
     if not solver_args:
-        solver_args = {'max_iters': 100, 'tolerance': 1e-6}
-
-    solver_ctr = getattr(pgb.solver, solver)
-    # sparse = pgb.solver.gmres(executor, sparse_matrix, iter, reset, stop)
+        solver_args = {
+            "type": "solver::Gmres",
+            "preconditioner": {
+                "type": "preconditioner::Ilu",
+                "l_solver_type": "solver::LowerTrs",
+                "reverse_apply": False,
+                "factorization": {"type": "factorization::ParIlu"},
+            },
+            "criteria": [
+                {"type": "Iteration", "max_iters": 1000},
+                {"type": "ResidualNorm", "reduction_factor": 1e-7},
+            ],
+        }
     solver_executor = A.get_executor()
-    solver_inst = solver_ctr(solver_executor, A, **solver_args)
 
     if not initial_guess:
         initial_guess = pgb.matrix.dense(b.get_executor(), (b.dim[0], 1))
         initial_guess.fill(0.0)
 
-    solver_inst.apply(b, initial_guess)
-    return initial_guess
+    logger = pGB.solver.config_solve(
+        solver_executor, A, b, initial_guess, json.dumps(solver_args)
+    )
+    return logger, initial_guess