+++ tests/cpp_bindings/test_array.py
@@ -7,7 +7,7 @@ import sys
 import pytest
 import numpy as np
 
-sys.path.append("../build")
+sys.path.append("../../")
 import pyGinkgoBindings as pGB