+++ tests/cpp_bindings/test_cuda_executor.py
@@ -6,7 +6,7 @@ import pytest
 import sys
 import numpy as np
 
-sys.path.append("../build")
+sys.path.append("../../")
 import pyGinkgoBindings as pGB