+++ tests/cpp_bindings/test_solver.py
@@ -4,14 +4,14 @@
 
 import os
 import sys
-import pyGinkgoBindings as pgb
 import pytest
 
-sys.path.append("../build")
+sys.path.append("../../")
+import pyGinkgoBindings as pgb
 
 
 @pytest.mark.parametrize("solver_name", ["gmres"])
-class TestSparseMatrix:
+class TestSolverBinding:
     ref = pgb.ReferenceExecutor()
 
     solver_args = {